requirements.txt
openai
bs4
fastapi
lxml
requests
uvicorn[standard]
psycopg2
datetime
google-cloud-bigquery
pandas
db-dtypes
openai